requirements.txt
# requirements.txt
Flask
python-dotenv
playwright
gunicorn
httpx
Flask[async]